refactor: command output capturing order
I was having trouble with the old implementation, and I managed to
narrow it down to it having something to do with the order of the logic.
When running all the commands individually they seemed to work just fine
but not when launched from a script and I think it boils down to
launching the new window before capturing the output. While `-t !`
should work it simply didn't. So in order to remove one possible race
condition, I made the .tmux plugin to run the script via run-shell,
capture the content there, format the content, and only then create a
new window for showing the end result.